feat: improve support
