+++ packages/openapi-code-generator/package.json
@@ -27,10 +27,7 @@
     "@types/js-yaml": "^4.0.9",
     "@types/lodash": "^4.17.0",
     "@typespec/compiler": "^0.55.0",
-    "@typespec/http": "^0.55.0",
-    "@typespec/openapi": "^0.55.0",
     "@typespec/openapi3": "^0.55.0",
-    "@typespec/versioning": "^0.55.0",
     "typescript": "~5.3.3"
   },
   "dependencies": {
@@ -50,10 +47,7 @@
   },
   "optionalDependencies": {
     "@typespec/compiler": "^0.55.0",
-    "@typespec/http": "^0.55.0",
-    "@typespec/openapi": "^0.55.0",
-    "@typespec/openapi3": "^0.55.0",
-    "@typespec/versioning": "^0.55.0"
+    "@typespec/openapi3": "^0.55.0"
   },
   "files": [
     "src",
