+++ packages/openapi-code-generator/src/core/loaders/utils.ts
@@ -1,0 +1,3 @@
+export function isRemote(location: string): boolean {
+  return location.startsWith("http://") || location.startsWith("https://")
+}
