+++ packages/openapi-code-generator/src/core/schemas/tsconfig.schema.ts
