+++ packages/openapi-code-generator/src/index.ts
@@ -5,8 +5,8 @@ import "source-map-support/register"
 import path from "path"
 
 import {Command, Option} from "@commander-js/extra-typings"
-import {loadTsConfigCompilerOptions} from "./core/file-loader"
 import {Input, OperationGroupStrategy} from "./core/input"
+import {loadTsConfigCompilerOptions} from "./core/loaders/tsconfig.loader"
 import {logger} from "./core/logger"
 import {OpenapiLoader} from "./core/openapi-loader"
 import {OpenapiValidator} from "./core/openapi-validator"
